+++ packages/dart/lib/src/telemetry/processing/processor.dart
@@ -14,6 +14,9 @@ abstract class TelemetryProcessor {
   /// Adds a log to be processed and buffered.
   void addLog(SentryLog log);
 
+  /// Adds a metric to be processed and buffered.
+  void addMetric(SentryMetric metric);
+
   /// Flushes all buffered telemetry data.
   ///
   /// Returns a [Future] if any buffer performs async flushing, otherwise
@@ -27,6 +30,12 @@ abstract class TelemetryProcessor {
 /// instances. If no buffer is registered for a telemetry type, items are
 /// dropped with a warning.
 class DefaultTelemetryProcessor implements TelemetryProcessor {
+  /// The buffer for metric data, or `null` if metric buffering is disabled.
+  final TelemetryBuffer<SentryMetric>? _metricBuffer;
+
+  @visibleForTesting
+  TelemetryBuffer<SentryMetric>? get metricBuffer => _metricBuffer;
+
   /// The buffer for log data, or `null` if log buffering is disabled.
   final TelemetryBuffer<SentryLog>? _logBuffer;
 
@@ -34,38 +43,62 @@ class DefaultTelemetryProcessor implements TelemetryProcessor {
   TelemetryBuffer<SentryLog>? get logBuffer => _logBuffer;
 
   DefaultTelemetryProcessor({
+    TelemetryBuffer<SentryMetric>? metricBuffer,
     TelemetryBuffer<SentryLog>? logBuffer,
-  }) : _logBuffer = logBuffer;
+  })  : _metricBuffer = metricBuffer,
+        _logBuffer = logBuffer;
 
   @override
   void addLog(SentryLog log) {
-    if (logBuffer == null) {
+    if (_logBuffer == null) {
       internalLogger.warning(
         '$runtimeType: No buffer registered for ${log.runtimeType} - item was dropped',
       );
       return;
     }
 
-    logBuffer!.add(log);
+    _logBuffer.add(log);
+
+    internalLogger.debug(() =>
+        '$runtimeType: Log "${log.body}" (${log.level.name}) added to buffer');
+  }
+
+  @override
+  void addMetric(SentryMetric metric) {
+    if (_metricBuffer == null) {
+      internalLogger.warning(
+        '$runtimeType: No buffer registered for ${metric.runtimeType} - item was dropped',
+      );
+      return;
+    }
+
+    _metricBuffer.add(metric);
+
+    internalLogger.debug(() =>
+        '$runtimeType: Metric "${metric.name}" (${metric.value}) added to buffer');
   }
 
   @override
   FutureOr<void> flush() {
     internalLogger.debug('$runtimeType: Clearing buffers');
 
-    final result = logBuffer?.flush();
+    final results = [_logBuffer?.flush(), _metricBuffer?.flush()];
 
-    if (result is Future) {
-      return result;
+    final futures = results.whereType<Future>().toList();
+    if (futures.isEmpty) {
+      return null;
     }
 
-    return null;
+    return Future.wait(futures).then((_) {});
   }
 }
 
 class NoOpTelemetryProcessor implements TelemetryProcessor {
   @override
   void addLog(SentryLog log) {}
+
+  @override
+  void addMetric(SentryMetric log) {}
 
   @override
   FutureOr<void> flush() {}