+++ packages/dart/lib/src/telemetry/processing/processor_integration.dart
@@ -20,8 +20,9 @@ class InMemoryTelemetryProcessorIntegration extends Integration<SentryOptions> {
       return;
     }
 
-    options.telemetryProcessor =
-        DefaultTelemetryProcessor(logBuffer: _createLogBuffer(options));
+    options.telemetryProcessor = DefaultTelemetryProcessor(
+        logBuffer: _createLogBuffer(options),
+        metricBuffer: _createMetricBuffer(options));
 
     options.sdk.addIntegration(integrationName);
   }
@@ -31,6 +32,17 @@ class InMemoryTelemetryProcessorIntegration extends Integration<SentryOptions> {
           encoder: (SentryLog item) => utf8JsonEncoder.convert(item.toJson()),
           onFlush: (items) {
             final envelope = SentryEnvelope.fromLogsData(
+                items.map((item) => item).toList(), options.sdk);
+            return options.transport.send(envelope).then((_) {});
+          });
+
+  InMemoryTelemetryBuffer<SentryMetric> _createMetricBuffer(
+          SentryOptions options) =>
+      InMemoryTelemetryBuffer(
+          encoder: (SentryMetric item) =>
+              utf8JsonEncoder.convert(item.toJson()),
+          onFlush: (items) {
+            final envelope = SentryEnvelope.fromMetricsData(
                 items.map((item) => item).toList(), options.sdk);
             return options.transport.send(envelope).then((_) {});
           });